Get the whole portable source set compiling: 386/386
Was 352/397. The 140 errors collapsed into far fewer real problems than
the count suggested.

sysNew.h: implement stack_new for non-CodeWarrior compilers. The #else
branch was a placeholder commented "This must be replaced with something
legal or be removed", and the note above it already described the
intended semantics -- placement new on stack-allocated space -- so it now
expands to exactly that over __builtin_alloca. That one macro accounted
for 116 of the 140 errors across ~30 files. The __MWERKS__ branch is
untouched, so the matching build is unaffected.

Two of the pointer-to-int casts were genuine 64-bit defects rather than
cosmetic truncation:

  objectMgr.cpp round-tripped mObjectPool through int before doing
  element arithmetic, so every address past the first was garbage on a
  64-bit host. mObjectPool is u8*; plain pointer arithmetic is correct
  and simpler.

  MenuItem::mData held a StageInfo* in an int that mapSelect.cpp casts
  back to a pointer. mData and the Menu API now use sptr, added to
  types.h. sptr is 32-bit under the PowerPC EABI, so struct layout and
  the matching build do not change.

The remaining nine casts are AgeServer telemetry and one debug PRINT,
where pointers are opaque handles on a 32-bit wire protocol and nothing
reads them back as addresses. Truncation there is now explicit.

ogMenu.cpp's bool draw() returns nothing, faithfully to the original;
the original is kept under __MWERKS__ and other compilers get a value.
node.cpp repeated a default argument on a constructor definition.

The 11 src/sysCore files needing windows.h or gl/gl.h are excluded via
tools/port/excluded.txt rather than shimmed -- they are Nintendo's Win32
dev tooling, and faking the headers would invent coverage we don't have.
oglGraphics.cpp is listed there too, to return once there is a GL target.

baseline.txt is now empty: any failure is a regression.

+++ src/plugPikiKando/objectMgr.cpp
@@ -749,7 +749,9 @@ int PolyObjectMgr::getTemplateIndex(int id)
  */
 Creature* PolyObjectMgr::get(int i)
 {
-	return (Creature*)((int)mObjectPool + mMaxSize * i);
+	// mObjectPool is u8*, so this is already byte-wise arithmetic. The original
+	// round-tripped it through int, which truncates a 64-bit pointer.
+	return (Creature*)(mObjectPool + mMaxSize * i);
 }
 
 /**
